browser-use-simple/pyproject.toml
[project]
name = "browser-use-simple"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "browser-use",
    "python-dotenv",
]
